src/app/questions/mortgage-intent/refinance/core/loan-amount/loan-amount.guard.spec.ts
import {
  TestBed,
  inject
} from '@angular/core/testing';

import { CanActivateQuestionsMortgageIntentLoanAmount } from './loan-amount.guard';
import { MortgageIntentRefinanceRoutingService } from '../refinance-routing.service';

describe('CanActivateQuestionsMortgageIntentLoanAmount', () => {
  let guard;
  let service;
  beforeEach(() => {
    TestBed.configureTestingModule({
      providers: [
        CanActivateQuestionsMortgageIntentLoanAmount,
        {
          provide: MortgageIntentRefinanceRoutingService,
          useValue: jasmine.createSpyObj('routingService', ['validLoanAmount'])
        }
      ]
    });
  });

  beforeEach(inject([MortgageIntentRefinanceRoutingService], (_service_) => {
    service = _service_;
  }));

  beforeEach(inject([CanActivateQuestionsMortgageIntentLoanAmount], (_guard_) => {
    guard = _guard_;
  }));

  it('should create', () => {
    expect(guard).toBeTruthy();
  });

  it('should have the canActivate method', () => {
    expect(guard.canActivate).toEqual(jasmine.any(Function));
  });

  describe('canActivate', () => {
    it('should return routingService.validLoanAmount', () => {
      service.validLoanAmount.and.returnValue(Promise.resolve(true));
      guard.canActivate().then(bool => {
        expect(bool).toEqual(true);
        expect(service.validLoanAmount).toHaveBeenCalledWith();
        expect(service.validLoanAmount).toHaveBeenCalledTimes(1);
      });

      service.validLoanAmount.calls.reset();
      service.validLoanAmount.and.returnValue(Promise.resolve(false));
      guard.canActivate().then(bool => {
        expect(bool).toEqual(false);
        expect(service.validLoanAmount).toHaveBeenCalledWith();
        expect(service.validLoanAmount).toHaveBeenCalledTimes(1);
      });
    });
  });
});
